Big structure change plus checks for QAV bandwidth.
Moved the test definitions to a configs sub-folder. Added the actions, graph, state_rendering and path_setup modules to try to keep the code clean.
