+++ tests/auto_test/configs/basics/check_connections.json
