+++ tests/auto_test/configs/basics/config.json
