+++ tests/auto_test/configs/regressions/bug_15151.json
