+++ tests/auto_test/configs/regressions/bug_15174.json
