+++ tests/auto_test/configs/regressions/bug_15175.json
