+++ tests/auto_test/configs/regressions/bug_15186.json
