+++ tests/auto_test/path_setup.py
@@ -1,0 +1,11 @@
+import os
+import sys
+
+def get_parent(full_path):
+  (parent, file) = os.path.split(full_path)
+  return parent
+
+# Configure the path so that the test framework will be found
+rootDir = get_parent(get_parent(get_parent(get_parent(os.path.realpath(__file__)))))
+sys.path.append(os.path.join(rootDir,'test_framework'))
+
